agregar peticiones preguntas
