+++ lib/src/models/question_model.dart
@@ -4,26 +4,32 @@
 
 import 'dart:convert';
 
-QuestionModel questionModelFromJson(String str) => QuestionModel.fromJson(json.decode(str));
+QuestionModel questionModelFromJson(String str) =>
+    QuestionModel.fromJson(json.decode(str));
 
 String questionModelToJson(QuestionModel data) => json.encode(data.toJson());
 
 class QuestionModel {
-    int id;
-    String ques;
+  int id;
+  String ques;
+  String type;
 
-    QuestionModel({
-        this.id,
-        this.ques,
-    });
+  QuestionModel({
+    this.id,
+    this.ques,
+    this.type
+  });
 
-    factory QuestionModel.fromJson(Map<String, dynamic> json) => new QuestionModel(
-        id: json["id"],
-        ques: json["ques"],
-    );
+  factory QuestionModel.fromJson(Map<String, dynamic> json) =>
+      new QuestionModel(
+        id: json["idpregunta"],
+        ques: json["pregunta"],
+        type: json["tipo"]
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "ques": ques,
-    };
+        "type": type
+      };
 }
